feat: ListActivity.kt 서버 통신 연결 및 HomeActivity.kt 으로부터 name, thumbnail 정보 추가로 받아 사용

+++ app/src/main/java/com/sieun/supporters_android/api/ApiRepository.kt
@@ -9,6 +9,6 @@ class ApiRepository @Inject constructor(
 ) {
     suspend fun getHome() = withContext(Dispatchers.IO) { service.getHome() }
 
-    suspend fun fetchCategoryItem(categoryNum: Int) = service.fetchCategoryItem(categoryNum)
+    suspend fun fetchCategoryItem(categoryNum: Long) = service.fetchCategoryItem(categoryNum)
 
 }
